requirements.txt
future==0.18.2
numpy==1.18.4
Pillow==7.1.2
PyYAML==5.3.1
scipy==1.4.1
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
